RavenDB-17070 Using Corax for AutoIndexes.

+++ src/Raven.Server/Documents/Indexes/Persistence/IndexReadOperationBase.cs
@@ -25,7 +25,7 @@ namespace Raven.Server.Documents.Indexes.Persistence
         {
         }
 
-        public abstract int EntriesCount();
+        public abstract long EntriesCount();
 
         public abstract IEnumerable<QueryResult> Query(IndexQueryServerSide query, QueryTimingsScope queryTimings, FieldsToFetch fieldsToFetch,
             Reference<int> totalResults, Reference<int> skippedResults, IQueryResultRetriever retriever, DocumentsOperationContext documentsContext,
